requirements.txt
requests
Pillow
transformers
python-telegram-bot===20.0b0
openai
python-dotenv
torch